Remove playground test code and add e2e test for orientation change on InformationPanel.
